fix(cc-components): move outdial into task folder
