+++ packages/contact-center/cc-components/src/components/StationLogin/station-login.types.ts
@@ -95,7 +95,7 @@ export interface IStationLoginProps {
   showMultipleLoginAlert: boolean;
 }
 
-export type StationLoginPresentationalProps = Pick<
+export type StationLoginComponentProps = Pick<
   IStationLoginProps,
   | 'teams'
   | 'loginOptions'
