+++ packages/contact-center/cc-components/src/components/task/CallControl/call-control.tsx
@@ -1,12 +1,12 @@
 import React, {useEffect, useState} from 'react';
 
-import {CallControlPresentationalProps} from '../task.types';
+import {CallControlComponentProps} from '../task.types';
 import './call-control.styles.scss';
 import {PopoverNext, SelectNext, TooltipNext, Text, ButtonCircle, ButtonPill} from '@momentum-ui/react-collaboration';
 import {Item} from '@react-stately/collections';
 import {Icon} from '@momentum-design/components/dist/react';
 
-function CallControlComponent(props: CallControlPresentationalProps) {
+function CallControlComponent(props: CallControlComponentProps) {
   const [selectedWrapupReason, setSelectedWrapupReason] = useState<string | null>(null);
   const [selectedWrapupId, setSelectedWrapupId] = useState<string | null>(null);
