+++ packages/contact-center/cc-components/src/components/task/OutdialCall/out-dial-call.styles.scss
