+++ packages/contact-center/cc-components/src/components/task/TaskList/task-list.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import {TaskListPresentationalProps} from '../task.types';
+import {TaskListComponentProps} from '../task.types';
 import Task from '../Task';
 import './styles.scss';
 
-const TaskListComponent: React.FunctionComponent<TaskListPresentationalProps> = (props) => {
+const TaskListComponent: React.FunctionComponent<TaskListComponentProps> = (props) => {
   const {currentTask, taskList, acceptTask, declineTask, isBrowser} = props;
   if (taskList.length <= 0) {
     return <></>; // hidden component
