+++ packages/contact-center/cc-components/src/wc.ts
@@ -4,7 +4,7 @@ import StationLoginComponent from './components/StationLogin/station-login';
 import CallControlComponent from './components/task/CallControl/call-control';
 import IncomingTaskComponent from './components/task/IncomingTask/incoming-task';
 import TaskListComponent from './components/task/TaskList/task-list';
-import OutdialCallComponent from './components/OutdialCall/out-dial-call';
+import OutdialCallComponent from './components/task/OutdialCall/out-dial-call';
 
 const WebUserState = r2wc(UserStateComponent, {
   props: {
